Add nightly for experimental Konversation Qt Quick branch
Summary:
I'd like to add Konversation's wip/qtquick branch to the Flatpak
repo for collaboration with designers.

This is my attempt at a recipe. I'm not sure if this is an
appropriate way to do it. I also don't know if conflicting
files are an issue, or how Discover would handle the two
packages co-existing considering they have the same .desktop
file and appstream data.

Hoping for feedback.

Reviewers: apol

Differential Revision: https://phabricator.kde.org/D7649